修改bug,完成所有功能

+++ lib/my_home/my_home_logic.dart
@@ -1,14 +1,13 @@
-import 'dart:convert';
-import 'dart:ffi';
 import 'dart:math';
-import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:sqflite/sqflite.dart';
 import 'package:ycd/my_db/DbHelper.dart';
 import 'package:ycd/my_db/Table1Model.dart';
+import 'package:ycd/utils/loading.dart';
 import '../my_db/Table2Model.dart';
-import '../utils/my_character.dart';
 import 'my_home_state.dart';
+import 'my_home_view.dart';
 
 class MyHomeLogic extends GetxController {
   final MyState state = MyState();
@@ -17,15 +16,75 @@ class MyHomeLogic extends GetxController {
   final scrollController = ScrollController();
   final textEditingController = TextEditingController();
 
+  FixedExtentScrollController? fixedExtentScrollController;
+
   @override
   void onInit() {
     super.onInit();
     _instance = DbHelper.instance.getDb();
 
     List.generate(32, (index) => state.totalValue.add('$index'));
-    state.chartData.value = List.generate(50, (index) => SalesData(index, Random().nextInt(1).toDouble())).toList();
+
     queryAll();
-    textEditingController.addListener(() => state.bettingMoney = textEditingController.text);
+    textEditingController.addListener(
+      () {
+        state.bettingMoney = textEditingController.text;
+        if (textEditingController.text.isNotEmpty) {
+          ///总体
+          state.totalValue[20] = pVal1();
+
+          ///局部
+          state.totalValue[24] = pVal2();
+        }
+      },
+    );
+  }
+
+  showFunctionTypesAlert() {
+    fixedExtentScrollController = FixedExtentScrollController(initialItem: state.selectIndex.value);
+    Get.bottomSheet(const SinglePicker());
+  }
+
+  String pVal2() {
+    var x = double.parse(state.totalValue[18]); //总输赢
+    var y = double.parse(state.bettingMoney); //输入框下注额
+    var z = double.parse(removeChineseCharacters(state.totalValue[14])); //净胜
+    var c11 = (double.parse(removeChineseCharacters(state.totalValue[14])).abs()); //净胜绝对值
+    if (z == 0) {
+      return "";
+    } else if (z > 0) /*赢>输的情况*/ {
+      if ((c11 - 1) <= 0) {
+        return '${((x + y) / (c11 + 1)).toStringAsFixed(1)}/';
+      }
+      return '${((x + y) / (c11 + 1)).toStringAsFixed(1)}/${((x - y) / (c11 - 1)).toStringAsFixed(1)}';
+    } else {
+      if ((c11 - 1) <= 0) {
+        Loading.showToast(toast: '/${((x - y) / (c11 + 1)).toStringAsFixed(1)}');
+        return '/${((x - y) / (c11 + 1)).toStringAsFixed(1)}';
+      }
+      return '${((x + y) / (c11 - 1)).toStringAsFixed(1)}/${((x - y) / (c11 + 1)).toStringAsFixed(1)}';
+    }
+  }
+
+  String pVal1() {
+    var x = double.parse(state.totalValue[17]); //总输赢
+    var y = double.parse(state.bettingMoney); //输入框下注额
+    var z = double.parse(removeChineseCharacters(state.totalValue[13])); //净胜
+    var c11 = (double.parse(removeChineseCharacters(state.totalValue[13])).abs()); //净胜绝对值
+    if (z == 0) {
+      return "";
+    } else if (z > 0) /*赢>输的情况*/ {
+      if ((c11 - 1) <= 0) {
+        return '${((x + y) / (c11 + 1)).toStringAsFixed(1)}/';
+      }
+      return '${((x + y) / (c11 + 1)).toStringAsFixed(1)}/${((x - y) / (c11 - 1)).toStringAsFixed(1)}';
+    } else {
+      if ((c11 - 1) <= 0) {
+        Loading.showToast(toast: '/${((x - y) / (c11 + 1)).toStringAsFixed(1)}');
+        return '/${((x - y) / (c11 + 1)).toStringAsFixed(1)}';
+      }
+      return '${((x + y) / (c11 - 1)).toStringAsFixed(1)}/${((x - y) / (c11 + 1)).toStringAsFixed(1)}';
+    }
   }
 
   @override
@@ -40,7 +99,6 @@ class MyHomeLogic extends GetxController {
     }
     state.isCanPress = false;
     state.js2 = state.js2 + 1;
-    print(state.js2);
     state.totalValue[28] = "${state.js1}/${state.js2}";
     if (next(1, 90485) > 44625 - MyState.OFFSET8431) {
       state.totalValue[30] = '庄';
@@ -59,9 +117,11 @@ class MyHomeLogic extends GetxController {
       db.query(DbHelper.table1).then((value) {
         state.table1List.clear();
         for (var data in value) {
-          print(data);
           state.table1List.add(Table1Model.fromJson(data));
         }
+        state.totalValue[0] = '${state.table1List.last.columnBenjin}'; //本金
+        state.totalValue[19] = '${state.table1List.last.columnMean}'; //期望值
+        state.chartData.value = List.generate(50, (index) => SalesData(index, double.parse(state.table1List.last.columnBenjin.toString()))).toList();
         _queryAllTable2();
       });
     });
@@ -69,16 +129,20 @@ class MyHomeLogic extends GetxController {
 
   add(int i, String tableName, {Table1Model? table1, Table2Model? table2}) {
     if (state.randomValue.isEmpty) {
-      Get.snackbar("温馨提示", '请摇塞子', duration: const Duration(seconds: 2), snackPosition: SnackPosition.TOP);
+      Get.snackbar("温馨提示", '请摇塞子', duration: const Duration(seconds: 2), snackPosition: SnackPosition.TOP, backgroundColor: Colors.white.withOpacity(0.7));
       return;
     }
-    print('下注金额${state.bettingMoney}');
+
     if (state.bettingMoney.isEmpty) {
-      Get.snackbar("温馨提示", '请输入下注金额', duration: const Duration(seconds: 2), snackPosition: SnackPosition.TOP);
+      Get.snackbar("温馨提示", '请输入下注金额', duration: const Duration(seconds: 2), snackPosition: SnackPosition.TOP, backgroundColor: Colors.white.withOpacity(0.7));
+      return;
+    }
+    if (!state.bettingMoney.isNum) {
+      Get.snackbar("温馨提示", '请输入数字', duration: const Duration(seconds: 2), snackPosition: SnackPosition.TOP, backgroundColor: Colors.white.withOpacity(0.7));
       return;
     }
     if (!state.isCanPress) {
-      Get.snackbar("温馨提示", '速度太快', duration: const Duration(seconds: 2), snackPosition: SnackPosition.BOTTOM);
+      Get.snackbar("温馨提示", '速度太快', duration: const Duration(seconds: 2), snackPosition: SnackPosition.BOTTOM, backgroundColor: Colors.white.withOpacity(0.7));
       return;
     }
     state.isCanPress = false;
@@ -110,41 +174,41 @@ class MyHomeLogic extends GetxController {
         state.table2List.clear();
         for (var data in value) {
           state.table2List.add(Table2Model.fromJson(data));
-          print(data);
         }
         scrollController.jumpTo(scrollController.position.maxScrollExtent + 50);
-        //计算
-        statisticalArea();
+        if (state.table2List.isNotEmpty) {
+          //计算
+          statisticalArea();
+        } else {
+          state.isCanPress = true;
+        }
       });
     });
   }
 
   getCharts() {
-    double current = 0 /* double.parse(state.totalValue[0])*/;
     var listZiJi = <double>[];
     var c = 0;
     for (var i = 0; i < state.table2List.length; i++) {
-      current += double.parse(state.table2List[i].colmunShuyingzhi.toString());
       c++;
       if (c > state.chartData.length) {
-        return;
+        break;
       }
-      listZiJi.add(current);
+      listZiJi.add(double.parse(state.table2List[i].columnCurrentJin.toString()));
     }
-    print('测试 原始资金表：${listZiJi}');
-    var z = 0;
-    for (var i = listZiJi.length - 1; i >= 0; i--) {
-      z++;
-      state.chartData[state.chartData.length - z].sales = listZiJi[i];
+    if (listZiJi.isNotEmpty) {
+      print(listZiJi);
+      var z = 0;
+      for (var i = listZiJi.length - 1; i >= 0; i--) {
+        z++;
+        state.chartData[state.chartData.length - z].sales = listZiJi[i];
+      }
+      var removeLast = state.chartData.removeLast();
+      Future.delayed(const Duration(milliseconds: 300), () => state.chartData.add(removeLast));
     }
-    var removeLast = state.chartData.removeLast();
-    Future.delayed(const Duration(milliseconds: 300), () => state.chartData.add(removeLast));
   }
 
   void statisticalArea() {
-    state.totalValue[0] = '${state.table1List.last.columnBenjin}'; //本金
-    state.totalValue[19] = '${state.table1List.last.columnMean}'; //期望值
-
     //图表区
     getCharts();
 
@@ -164,18 +228,15 @@ class MyHomeLogic extends GetxController {
         zt_y++;
       }
     }
-    state.totalValue[5] = '${zt_y}'; //胜
+    state.totalValue[5] = '$zt_y'; //胜
     state.totalValue[9] = '${(zt_y / double.parse(state.totalValue[1]) * 100).toStringAsFixed(2)}%'; //胜率
     state.totalValue[13] = '${zt_y.abs() - zt_s.abs()}'; //净胜~须多少手回到50%
-    state.totalValue[17] = '${zt_syz}'; //一共输赢多少钱
-    state.totalValue[21] = '${zt_syz / double.parse(removeChineseCharacters(state.totalValue[13]))}'; //平均输赢多少钱
+    state.totalValue[17] = '$zt_syz'; //一共输赢多少钱
+    state.totalValue[21] = (zt_syz / double.parse(removeChineseCharacters(state.totalValue[13])).abs()).toStringAsFixed(2); //平均输赢多少钱
     var d = (double.parse(state.totalValue[1]) + 1) * double.parse(state.totalValue[19]); //期望一共的值
     var parse = int.parse(state.totalValue[13]).abs();
-    state.totalValue[25] = zt_syz < 0
-        ? /*state.randomValue == '闲' || state.randomValue.isEmpty
-            ?*/
-        '须${((zt_syz.abs() + d) / parse).toStringAsFixed(1)}x${parse}'
-        : '须${((zt_syz.abs() + d) / double.parse(state.totalValue[23]) / parse).toStringAsFixed(1)}x${parse}' /*: '可负${((zt_syz.abs() - d) / parse).toStringAsFixed(1)}x${parse}'*/; //还需，可负
+    state.totalValue[25] =
+        zt_syz < 0 ? '须${((zt_syz.abs() + d) / parse).toStringAsFixed(1)}x$parse' : '可负${((zt_syz.abs() - d) / parse).toStringAsFixed(1)}x$parse'; //还需，可负
     state.totalValue[29] = '${state.table1List.last.columnRestartIndex}'; //流水索引 重启位置
 
     state.totalValue[4] = "${double.parse(state.totalValue[0]) + zt_syz}"; //当前金额
@@ -199,39 +260,59 @@ class MyHomeLogic extends GetxController {
       }
     }
     state.totalValue[6] = '$jb_y'; //净胜
-    state.totalValue[10] = '${(jb_y / jb_count * 100).toStringAsFixed(2)}%'; //胜率
+    state.totalValue[10] = jb_count == 0 ? "" : '${(jb_y / jb_count * 100).toStringAsFixed(2)}%'; //胜率
     state.totalValue[14] = '${jb_y.abs() - jb_s.abs()}'; //净胜~须多少手回到50%
     state.totalValue[18] = '$jb_syz'; //一共输赢多少钱
-    state.totalValue[22] = '${jb_syz / double.parse(removeChineseCharacters(state.totalValue[14]))}'; //平均输赢多少钱
+    state.totalValue[22] = jb_count <= 0 ? "" : (jb_syz / double.parse(removeChineseCharacters(state.totalValue[14])).abs()).toStringAsFixed(3); //平均赢
     var dJ = (jb_count + 1) * double.parse(state.totalValue[19]); //期望一共的值
     parse = int.parse(state.totalValue[14]).abs();
-    state.totalValue[26] = /*state.randomValue == '闲' || state.randomValue.isEmpty
-        ? */
-            jb_syz < 0 ? '须${((jb_syz.abs() + dJ) / parse).toStringAsFixed(1)}x${parse}' : '可负${((jb_syz.abs() - dJ) / parse).toStringAsFixed(1)}x${parse}'
-        /* : jb_syz < 0
-            ? '须${((jb_syz.abs() + dJ) / double.parse(state.totalValue[23]) / parse).toStringAsFixed(1)}x${parse}'
-            : '可负${((jb_syz.abs() - dJ) / parse).toStringAsFixed(1)}x${parse}'*/
-        ; //还需，可负
+    state.totalValue[26] = jb_syz < 0
+        ? parse == 0
+            ? ''
+            : '须${((jb_syz.abs() + dJ) / parse).toStringAsFixed(1)}x$parse'
+        : parse == 0
+            ? ''
+            : '可负${((jb_syz.abs() - dJ) / parse).toStringAsFixed(1)}x$parse';
 
     ///第四列
     state.totalValue[3] = '流水$runningWater';
-    state.totalValue[7] = '均利${zt_syz / state.table2List.length}';
+    state.totalValue[7] = '均利${zt_syz ~/ state.table2List.length}';
+    state.totalValue[11] = '连胜负';
     state.totalValue[23] = '${state.table1List.last.columnYongJin}'; //赔率
-    state.totalValue[27] = zt_syz>0?"": (double.parse(state.totalValue[21]) / double.parse(state.totalValue[23])).toStringAsFixed(2); //打庄需要
-    state.totalValue[31] = jb_syz>0?"":(double.parse(state.totalValue[22]) / double.parse(state.totalValue[23])).toStringAsFixed(2); //打庄需要
+    state.totalValue[27] = zt_syz > 0
+        ? ""
+        : double.parse(state.totalValue[21]).isNaN || !state.totalValue[25].split("x")[0].replaceAll("须", '').isNum
+            ? ""
+            : (double.parse(state.totalValue[25].split("x")[0].replaceAll("须", '')) / double.parse(state.totalValue[23])).toStringAsFixed(2); //打庄需要
+    state.totalValue[31] = jb_syz > 0
+        ? ""
+        : state.totalValue[22].isEmpty || double.parse(state.totalValue[22]).isNaN || state.totalValue[26].split("x")[0].replaceAll("须", '').isEmpty
+            ? ""
+            : (double.parse(state.totalValue[26].split("x")[0].replaceAll("须", '')) / double.parse(state.totalValue[23])).toStringAsFixed(2);
+
+    //预测平均值
+    if (textEditingController.text.isNotEmpty) {
+      ///总体
+      state.totalValue[20] = pVal1();
+
+      ///局部
+      state.totalValue[24] = pVal2();
+    }
     state.isCanPress = true;
-    print(state.isCanPress);
   }
 
-  getCurrentJin(int i, double parse) {
+  getCurrentJin(int i, double playMoney) {
+    var lastJinE = state.table2List.isEmpty
+        ? double.parse(state.table1List.last.columnBenjin.toString())
+        : double.parse(state.table2List.last.columnCurrentJin.toString());
     switch (i) {
       case 1:
-        return (double.parse(state.table1List.last.columnBenjin.toString()) + parse);
+        return (lastJinE + playMoney);
       case 2:
-        return (double.parse(state.table1List.last.columnBenjin.toString()) + parse * double.parse(state.totalValue[27]));
+        return (lastJinE) + playMoney * double.parse(state.totalValue[23]);
       case 3:
       case 4:
-        return (double.parse(state.table1List.last.columnBenjin.toString()) - parse);
+        return (lastJinE) - playMoney;
     }
   }
 
@@ -242,9 +323,7 @@ class MyHomeLogic extends GetxController {
       case 2: //庄赢
         double parse = double.parse(state.bettingMoney);
         var xx = parse * double.parse(state.totalValue[23]);
-        print('庄赢值=>$xx');
         String syz /*庄赢值*/ = xx.toStringAsFixed(2); //四舍五入保留两位小数
-        print('庄赢值$syz');
         return syz;
       case 3:
       case 4:
@@ -253,7 +332,9 @@ class MyHomeLogic extends GetxController {
   }
 
   void deleteLast() {
-    _instance?.then((db) => db.delete(DbHelper.table2, where: 'table2Id =?', whereArgs: [state.table2List.last.table2Id]).then((value) => queryAll()));
+    if (state.table2List.isNotEmpty) {
+      _instance?.then((db) => db.delete(DbHelper.table2, where: 'table2Id =?', whereArgs: [state.table2List.last.table2Id]).then((value) => queryAll()));
+    }
   }
 
   void updateSqlite(int index) {
@@ -275,6 +356,63 @@ class MyHomeLogic extends GetxController {
               columnLiushuiIndex: valueC.last['column_liushui_index'].toString(),
             ).toJson())
         .then((value) => queryAll()))));
+  }
+
+  void updateBenJin(String b) {
+    // valueC.last..update('column_restart_index', (value) => "${state.table2List.length - 1}")
+    _instance?.then((value) => value.query(DbHelper.table1).then((valueC) => _instance?.then((value) => value
+        .insert(
+            DbHelper.table1,
+            Table1Model(
+              columnBenjin: b,
+              columnYongJin: valueC.last['column_yongJin'].toString(),
+              columnMean: valueC.last['column_mean'].toString(),
+              columnRestartIndex: valueC.last['column_restart_index'].toString(),
+              columnLiushuiIndex: valueC.last['column_liushui_index'].toString(),
+            ).toJson())
+        .then((value) => queryAll()))));
+  }
+
+  void functionConfirm(int i) {
+    switch (i) {
+      case 0:
+        break;
+      case 1:
+        break;
+      case 2:
+        if (textEditingController.text.toString().isEmpty) {
+          Loading.showToast(toast: '请输入金额 ${textEditingController.text} ');
+          break;
+        }
+        if (!textEditingController.text.toString().isNum) {
+          Loading.showToast(toast: '请输入数字 ${textEditingController.text} ');
+          break;
+        }
+        updateBenJin(textEditingController.text.toString());
+        break;
+      case 3:
+        break;
+      case 4:
+        break;
+      case 5:
+        break;
+      case 6:
+        break;
+      case 7:
+        Get.defaultDialog(
+          barrierDismissible: false,
+          title: '警告',
+          content: const Text('是否重启'),
+          onCancel: () {},
+          onConfirm: () {
+            reStart();
+            Get.back();
+          },
+        );
+        break;
+      case 8:
+        break;
+    }
   }
 }
